Add BaseTemplate, Home template, 404 page, and lazy loading for images (#34)
* feat: Add BaseTemplate component

The `BaseTemplate` component is added to provide a base layout for the application. It includes a `Navbar` component at the top, followed by the main content wrapped in a `div`, and a `Footer` component at the bottom. This component will be used as a template for other pages in the application.

* feat: Add Home template component

The `Home` template component is added to provide the structure for the home page of the application. It includes different sections such as hero, about, mission, donations, and newsletter. This component will be used as a template for the home page content.

Co-authored-by: lucaoskaique <lucas.kaique175@hotmail.com>
Co-authored-by: Victor Feitosa Garcia <vfeitosa.garcia@gmail.com>

* feat: Add 404 page component and empty component for the page

* feat: Add lazy loading for images and update next-seo.config.js

The code changes include adding lazy loading for images to improve page load performance. Additionally, the next-seo.config.js file is updated to include Open Graph and Twitter metadata for better social media sharing.

Co-authored-by: lucaoskaique <lucas.kaique175@hotmail.com>
Co-authored-by: Victor Feitosa Garcia <vfeitosa.garcia@gmail.com>

* feat: Add lazy loading for images in Button component

The code changes include adding lazy loading for images in the Button component to improve page load performance.

Co-authored-by: lucaoskaique <lucas.kaique175@hotmail.com>
Co-authored-by: Victor Feitosa Garcia <vfeitosa.garcia@gmail.com>

* feat: Add lazy loading for images in Button component

* feat: Update Base template component

The `Base` template component is updated to improve the layout and structure of the application. Changes include:
- Adding a flex container with vertical column layout
- Updating the navbar and footer sections with appropriate styling
- Adjusting the main content section to take up remaining space

* feat: fixed types

* chore: Update CI workflow name to "CI"

---------

Co-authored-by: Victor Feitosa Garcia <vfeitosa.garcia@gmail.com>