Add functionality to generate sitemap.xml file
This commit adds a new file, generateSitemap.ts, which contains the logic to generate a sitemap.xml file for the blog. The sitemap includes both normal page routes and post routes. The function takes an array of posts as input and uses the globby library to find all the page routes. It then creates an XML string representing the sitemap and writes it to the public/sitemap.xml file.

Closes #123